LayoutTests/imported/w3c/web-platform-tests/svg/animations/animate-currentcolor-in-visited-link-expected.txt


FAIL SMIL animating 'currentcolor' on target within visited link assert_equals: currentcolor animation does not leak :visited expected "rgb(0, 0, 255)" but got "rgb(173, 216, 230)"
